regression-suite/requirements.txt
# Core
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.10.0

# Optional: For real model APIs
# openai>=1.0.0
# anthropic>=0.18.0

# Development
# pytest>=7.0.0
# black>=23.0.0
